+++ src/feeds/topPostsFeed.ts
@@ -23,8 +23,7 @@ export default async function topPostsFeed(api: mastodon.rest.Client): Promise<T
     // Get list of top mastodon servers // TODO: what does "top" mean here?
     const topServerDomains = Object.keys(coreServers)
                                    .filter(s => s !== "undefined" && typeof s !== "undefined" && s.length > 0)
-                                   .sort((a, b) => (coreServers[b] - coreServers[a]))  // TODO: wtf is this comparison?
-                                   .slice(0, NUM_SERVERS_TO_POLL);
+                                   .sort((a, b) => (coreServers[b] - coreServers[a]));  // TODO: wtf is this comparison?
 
     if (topServerDomains.length == 0) {
         console.warn("No mastodon servers found to get topPostsFeed data from!");