+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+`2024-11-27`
+* Add a list of default servers to check for trending toots if user isn't following very many accounts
+
 `2024-11-26`
 * Add `FollowedTagsScorer`